+++ Model/Config/File/PrivateKeyPem.php
@@ -1,0 +1,97 @@
+<?php
+/**
+ * Copyright © Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+namespace Amazon\Pay\Model\Config\File;
+
+/**
+ * Class PrivateKeyPem
+ * @package Amazon\Pay\Model\Config\File
+ */
+class PrivateKeyPem extends \Magento\Config\Model\Config\Backend\File
+{
+    /**
+     * @var \Magento\Framework\Encryption\EncryptorInterface
+     */
+    protected $encryptor;
+
+    /**
+     * PrivateKeyPem constructor.
+     * @param \Magento\Framework\Model\Context $context
+     * @param \Magento\Framework\Registry $registry
+     * @param \Magento\Framework\App\Config\ScopeConfigInterface $config
+     * @param \Magento\Framework\App\Cache\TypeListInterface $cacheTypeList
+     * @param \Magento\MediaStorage\Model\File\UploaderFactory $uploaderFactory
+     * @param \Magento\Config\Model\Config\Backend\File\RequestData\RequestDataInterface $requestData
+     * @param \Magento\Framework\Filesystem $filesystem
+     * @param \Magento\Framework\Encryption\EncryptorInterface $encryptor
+     * @param \Magento\Framework\Model\ResourceModel\AbstractResource|null $resource
+     * @param \Magento\Framework\Data\Collection\AbstractDb|null $resourceCollection
+     * @param array $data
+     */
+    public function __construct(
+        \Magento\Framework\Model\Context $context,
+        \Magento\Framework\Registry $registry,
+        \Magento\Framework\App\Config\ScopeConfigInterface $config,
+        \Magento\Framework\App\Cache\TypeListInterface $cacheTypeList,
+        \Magento\MediaStorage\Model\File\UploaderFactory $uploaderFactory,
+        \Magento\Config\Model\Config\Backend\File\RequestData\RequestDataInterface $requestData,
+        Filesystem $filesystem,
+        \Magento\Framework\Encryption\EncryptorInterface $encryptor,
+        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
+        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
+        array $data = []
+    )
+    {
+        parent::__construct($context, $registry, $config, $cacheTypeList, $uploaderFactory, $requestData, $filesystem, $resource, $resourceCollection, $data);
+
+        $this->encryptor = $encryptor;
+    }
+
+    /**
+     * @return $this|PrivateKeyPem
+     */
+    public function beforeSave()
+    {
+        $value = $this->getValue();
+        $file = $this->getFileData();
+        if (!empty($file)) {
+            try {
+                $uploader = $this->_uploaderFactory->create(['fileId' => $file]);
+                $uploader->setAllowedExtensions($this->_getAllowedExtensions());
+                $uploader->setAllowRenameFiles(true);
+                $uploader->addValidateCallback('size', $this, 'validateMaxSize');
+                if ($uploader->validateFile()) {
+                    $privateKey = file_get_contents($value['tmp_name']);
+                    $this->setValue($this->encryptor->encrypt($privateKey));
+                }
+            } catch (\Exception $e) {
+                throw new \Magento\Framework\Exception\LocalizedException(__('%1', $e->getMessage()));
+            }
+        }
+        else {
+            $this->unsValue();
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return array|string[]
+     */
+    public function _getAllowedExtensions()
+    {
+        return ['pem'];
+    }
+}